agent/pyproject.toml
[project]
name = "open-novel-agent"
version = "0.1.0"
description = "External LangGraph agent service for OpenNovel intent validation and narrative rendering."
requires-python = ">=3.9"
dependencies = [
  "chromadb>=1.0,<2.0",
  "fastapi>=0.115,<1.0",
  "httpx>=0.28,<1.0",
  "langgraph>=0.2,<1.0",
  "pydantic>=2.10,<3.0",
  "python-dotenv>=1.0,<2.0",
  "uvicorn>=0.34,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
